added authentication to user routers

+++ app/routers/user.py
@@ -1,7 +1,7 @@
 from fastapi import Response, status, HTTPException, Depends, APIRouter
 from sqlalchemy.orm import Session
 from typing import List
-from .. import models, schemas, utils
+from .. import models, schemas, utils, oauth2
 from ..database import get_db
 
 router = APIRouter(
@@ -11,7 +11,7 @@ router = APIRouter(
 
 # /users/
 # /users
-#create /get /get/{id} /post /delete{id} /put
+
 
 @router.post("/", status_code=status.HTTP_201_CREATED, response_model=schemas.UserOut)
 def create_user(user: schemas.UserCreate, db: Session = Depends(get_db)):
@@ -35,13 +35,13 @@ def create_user(user: schemas.UserCreate, db: Session = Depends(get_db)):
 
 
 @router.get('/', response_model=List[schemas.UserOut])
-def get_users(db: Session = Depends(get_db)):
+def get_users(db: Session = Depends(get_db), current_user: int = Depends(oauth2.get_current_user)):
     users = db.query(models.User).all()
     return users
 
 
 @router.get('/{id}', response_model=schemas.UserOut)
-def get_user(id: int, db: Session = Depends(get_db), ):
+def get_user(id: int, db: Session = Depends(get_db), current_user: int = Depends(oauth2.get_current_user)):
     user = db.query(models.User).filter(models.User.id == id).first()
     if not user:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
@@ -51,15 +51,18 @@ def get_user(id: int, db: Session = Depends(get_db), ):
 
 
 @router.delete('/{id}', status_code=status.HTTP_204_NO_CONTENT)
-def delete_user(id: int, db: Session = Depends(get_db)):
+def delete_user(id: int, db: Session = Depends(get_db), current_user: int = Depends(oauth2.get_current_user)):
     
-    # check if id == current_user.id
+    if id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, 
+        detail='Not authorized to perform requested action')
+
     user_query = db.query(models.User).filter(models.User.id == id)
     user = user_query.first()
 
     if not user:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
-                            detail=f"user with id: {id} does not exist")
+                            detail=f"User with id: {id} does not exist")
 
     user_query.delete()
     db.commit()
@@ -68,8 +71,12 @@ def delete_user(id: int, db: Session = Depends(get_db)):
 
 
 @router.put("/{id}", response_model=schemas.UserOut)
-def update_user(id: int, updated_user: schemas.UserCreate, db: Session = Depends(get_db)):
+def update_user(id: int, updated_user: schemas.UserCreate, db: Session = Depends(get_db), current_user: int = Depends(oauth2.get_current_user)):
 
+    if id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, 
+        detail='Not authorized to perform requested action')
+    
     user_query = db.query(models.User).filter(models.User.id == id)
 
     user = user_query.first()
@@ -79,9 +86,6 @@ def update_user(id: int, updated_user: schemas.UserCreate, db: Session = Depends
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail=f"user with id: {id} does not exist")
 
-    # if user.id != current_user.id:
-    #     raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
-    #                         detail="Not authorized to perform requested action")
 
     user_query.update(updated_user.dict(), synchronize_session=False)
